shuffle a bit
